ebiten, inpututil: add TouchPositionF and TouchPositionFInPreviousTick
CursorPositionF returns a high-precision cursor position, but touches
had no counterpart, so TouchPosition was limited to whole logical
pixels even where a platform reports a subpixel position.

The precision was discarded inside Ebitengine rather than unavailable:
internal/ui's Touch stored an integer position while the cursor was
already a float64. Every backend computed the logical position in
floating point and truncated it when building the Touch. Touch now
stores float64, and the conversion happens in the integer accessors.

inpututil gains TouchPositionFInPreviousTick beside
TouchPositionInPreviousTick, the only inpututil function carrying a
touch position.

On Android and iOS a position was truncated in Java and Objective-C
before reaching Go, so the mobile bridge now takes float64 as well.
This slightly changes what TouchPosition reports on Android, where a
position used to be truncated twice, once in pixels and once in
density independent pixels.

Nintendo Switch keeps its integer C struct, which is shared with an
implementation outside this repository. A position there is still
fractional after the device scale and screen scale division.

Closes #3504

+++ cmd/ebitenmobile/_files/EbitenView.java
@@ -156,10 +156,10 @@ public class EbitenView extends ViewGroup implements InputManager.InputDeviceLis
         int touchIndex = e.getActionIndex();
         for (int i = 0; i < e.getPointerCount(); i++) {
             int id = e.getPointerId(i);
-            int x = (int)e.getX(i);
-            int y = (int)e.getY(i);
+            double x = e.getX(i);
+            double y = e.getY(i);
             int action = (i == touchIndex) ? e.getActionMasked() : MotionEvent.ACTION_MOVE;
-            Ebitenmobileview.updateTouchesOnAndroid(action, id, (int)pxToDp(x), (int)pxToDp(y));
+            Ebitenmobileview.updateTouchesOnAndroid(action, id, pxToDp(x), pxToDp(y));
         }
         return true;
     }
